refactor(examples): split Examples shell into lazy chunk
Move the shell + 404 components into _base, lazy-load Examples only
in the top window, and keep main.tsx as a tiny branch: shell for the
parent frame, single App module for the iframe (selected via hash
key into the glob). Iframes no longer bundle the shell.
