+++ examples/_base/NotFound.tsx
@@ -1,0 +1,8 @@
+export function NotFound({ path }: { path: string; }) {
+  return (
+    <div style={{ padding: '4rem 2rem', textAlign: 'center' }}>
+      <h1>404</h1>
+      <p>No example matches <code>{path}</code>.</p>
+    </div>
+  );
+}
